Correct equirectangular map projection
BUGFIXes in coordiante conversion to local projection coordinates.
Correct max and min longitude and latitude calculation for viewRect
based on whether the projection should be rectangular (use the whole
viewrect) or bounded by a max and min longitude and latitude as in
Uranometria. The whole projection should then be inside (and only part)
of the viewport.